4.21 done

+++ part4/Bloglist/controllers/blogs.js
@@ -50,8 +50,25 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'Could not find a blog!!' })
+}
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+} else {
+    response.status(400).json({ error: 'incorrect token or/and incorrect user!!' })
+}
+
+
+
+if (!request.token || !decodedToken.id) {
+  return response.status(401).json({ error: 'add token or put valid token!!' })
+}
   })
